+++ app/Http/Controllers/UserController.php
@@ -4,8 +4,10 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 
+use Auth;
 use App\Http\Requests;
 use App\Http\Requests\UserRequest;
+use App\Http\Requests\ProfileRequest;
 
 class UserController extends Controller
 {
@@ -16,6 +18,7 @@ class UserController extends Controller
 	 */
 	public function create()
 	{
+		$this->authorize('create_user', \App\User::class);
 		$roles = \App\Role::all();
 
 		return View('user.create')->with('roles', $roles);
@@ -31,6 +34,9 @@ class UserController extends Controller
 	public function destroy($id, Request $request) 
 	{
 		$user = \App\User::findOrFail($id);
+		
+		$this->authorize('delete_user', $user);
+
 		$user->delete();
 
 		Session()->flash('flash_message', 'User successfully deleted!');
@@ -48,9 +54,28 @@ class UserController extends Controller
 	public function edit($id, Request $request) 
 	{
 		$user = \App\User::findOrFail($id);
+
+		$this->authorize('edit_user', $user);
+
 		$roles = \App\Role::all();
 
 		return View('user.edit')->with('user', $user)->with('roles', $roles)->with('checkedRoles', $user->roles);
+	}
+
+	/**
+	 * Show edit profile form.
+	 * 
+	 * @param  int  $id
+	 * @param  Request $request 
+	 * @return view
+	 */
+	public function editprofile(ProfileRequest $request) 
+	{
+		$user = \App\User::findOrFail(Auth::user()->id);
+
+		$this->authorize('edit_profile', \App\User::class);
+
+		return View('user.editprofile')->with('user', $user);
 	}
 
 	/**
@@ -60,6 +85,7 @@ class UserController extends Controller
 	 */
 	public function index() 
 	{
+		$this->authorize('list_users', \App\User::class);
 		$users = \App\User::latest()->paginate();
 
 		return View('user.index')->with('users', $users);
@@ -74,6 +100,7 @@ class UserController extends Controller
 	public function show($id)
 	{
 		$user = \App\User::findOrFail($id);
+		$this->authorize('show_user', $user);
 
 		return View('user.show')->with('user', $user);
 	}
@@ -86,6 +113,8 @@ class UserController extends Controller
 	 */
 	public function store(UserRequest $request) 
 	{
+		$this->authorize('create_user', \App\User::class);
+
 		$user = \App\User::create([
 								'name'		=> 	$request->input('name'),
 								'email'		=>	$request->input('email'),
@@ -110,6 +139,7 @@ class UserController extends Controller
 	public function update($id, UserRequest $request) 
 	{
 		$user = \App\User::findOrFail($id);
+		$this->authorize('edit_user', $user);
 
 		$data = [
 			'name'	=> 	$request->input('name'),
@@ -138,4 +168,32 @@ class UserController extends Controller
 
 		return redirect()->route('user.index');
 	}
+
+	/**
+	 * Save user profile info after update.
+	 * 
+	 * @param  Request 	$request
+	 * @return redirect
+	 */
+	public function updateprofile(ProfileRequest $request) 
+	{
+		$user = \App\User::findOrFail(Auth::user()->id);
+		$this->authorize('edit_profile', $user);
+
+		$data = [
+			'email'	=>	$request->input('email'),
+		];
+
+		// Only include password if we changed it,
+		// otherwise it will be set blank!
+		if (!empty($request->input('password'))) {
+			$data['password'] = bcrypt($request->input('password'));
+		}
+
+		$user->update($data);
+
+		Session()->flash('flash_message', 'Profile successfully updated!');
+
+		return redirect()->route('user.index');
+	}	
 }
