+++ resources/views/role/show.blade.php
@@ -10,14 +10,19 @@
 	<tr><th>ID</th><td><?php echo $role->id;?></td></tr>
 	<tr><th>Name</th><td><?php echo $role->name;?></td></tr>
 	<tr><th>Action</th>
-	<td><a href="{{ action('RoleController@edit', [$role->id]) }}"><button type="button" class="btn btn-xs btn-info">Edit</button></a> 
-	
-		{!! Form::open([
-			'route' => ['role.destroy', $role->id], 
-			'method' => 'DELETE'
-			]) !!}
-			{!! Form::submit('Delete', ['class' => 'btn btn-xs btn-danger', 'onclick' => 'return confirm(\'Are you sure?\')']) !!}
-		{!! Form::close() !!}
+	<td>
+		@can('edit_role', $role)
+			<a href="{{ action('RoleController@edit', [$role->id]) }}"><button type="button" class="btn btn-xs btn-info">Edit</button></a> 
+		@endcan
+
+		@can('delete_role', $role)
+			{!! Form::open([
+				'route' => ['role.destroy', $role->id], 
+				'method' => 'DELETE'
+				]) !!}
+				{!! Form::submit('Delete', ['class' => 'btn btn-xs btn-danger', 'onclick' => 'return confirm(\'Are you sure?\')']) !!}
+			{!! Form::close() !!}
+		@endcan
 	</td>
 	</tr>
 	</table>
